support non-200 resources: fix vgc tests
