+++ packages/visual-grid-client/test/fixtures/test.cdt.json
@@ -3,7 +3,7 @@
     "nodeType": 9,
     "childNodeIndexes": [
       1,
-      79
+      85
     ]
   },
   {
@@ -54,7 +54,7 @@
   },
   {
     "nodeType": 3,
-    "nodeValue": "\n    @import 'imported2.css';\n  "
+    "nodeValue": "@import url(\"imported2.css\");"
   },
   {
     "nodeType": 1,
@@ -79,6 +79,48 @@
       {
         "name": "href",
         "value": "https://fonts.googleapis.com/css?family=Zilla+Slab"
+      }
+    ],
+    "childNodeIndexes": []
+  },
+  {
+    "nodeType": 3,
+    "nodeValue": "\n  "
+  },
+  {
+    "nodeType": 1,
+    "nodeName": "LINK",
+    "attributes": [
+      {
+        "name": "rel",
+        "value": "stylesheet"
+      },
+      {
+        "name": "href",
+        "value": "/predefined-status/404"
+      }
+    ],
+    "childNodeIndexes": []
+  },
+  {
+    "nodeType": 3,
+    "nodeValue": " "
+  },
+  {
+    "nodeType": 3,
+    "nodeValue": "\n  "
+  },
+  {
+    "nodeType": 1,
+    "nodeName": "LINK",
+    "attributes": [
+      {
+        "name": "rel",
+        "value": "stylesheet"
+      },
+      {
+        "name": "href",
+        "value": "/predefined-status/500"
       }
     ],
     "childNodeIndexes": []
@@ -141,7 +183,12 @@
       11,
       12,
       13,
-      14
+      14,
+      15,
+      16,
+      17,
+      18,
+      19
     ]
   },
   {
@@ -166,7 +213,7 @@
       }
     ],
     "childNodeIndexes": [
-      18
+      23
     ]
   },
   {
@@ -187,7 +234,7 @@
       }
     ],
     "childNodeIndexes": [
-      21
+      26
     ]
   },
   {
@@ -212,7 +259,7 @@
       }
     ],
     "childNodeIndexes": [
-      24
+      29
     ]
   },
   {
@@ -237,7 +284,7 @@
       }
     ],
     "childNodeIndexes": [
-      27
+      32
     ]
   },
   {
@@ -253,7 +300,7 @@
     "nodeName": "DIV",
     "attributes": [],
     "childNodeIndexes": [
-      30
+      35
     ]
   },
   {
@@ -284,7 +331,7 @@
     "nodeName": "DIV",
     "attributes": [],
     "childNodeIndexes": [
-      35
+      40
     ]
   },
   {
@@ -319,7 +366,7 @@
     "nodeName": "DIV",
     "attributes": [],
     "childNodeIndexes": [
-      40
+      45
     ]
   },
   {
@@ -358,7 +405,7 @@
     "nodeName": "DIV",
     "attributes": [],
     "childNodeIndexes": [
-      46
+      51
     ]
   },
   {
@@ -406,7 +453,7 @@
       }
     ],
     "childNodeIndexes": [
-      53
+      58
     ]
   },
   {
@@ -423,8 +470,8 @@
       }
     ],
     "childNodeIndexes": [
-      52,
-      54
+      57,
+      59
     ]
   },
   {
@@ -449,7 +496,7 @@
       }
     ],
     "childNodeIndexes": [
-      57
+      62
     ]
   },
   {
@@ -474,7 +521,7 @@
       }
     ],
     "childNodeIndexes": [
-      60
+      65
     ]
   },
   {
@@ -499,7 +546,7 @@
       }
     ],
     "childNodeIndexes": [
-      63
+      68
     ]
   },
   {
@@ -524,7 +571,7 @@
       }
     ],
     "childNodeIndexes": [
-      66
+      71
     ]
   },
   {
@@ -541,17 +588,17 @@
       }
     ],
     "childNodeIndexes": [
-      51,
-      55,
       56,
-      58,
-      59,
+      60,
       61,
-      62,
+      63,
       64,
-      65,
+      66,
       67,
-      68
+      69,
+      70,
+      72,
+      73
     ]
   },
   {
@@ -568,13 +615,13 @@
       }
     ],
     "childNodeIndexes": [
-      45,
-      47,
-      48,
-      49,
       50,
-      69,
-      70
+      52,
+      53,
+      54,
+      55,
+      74,
+      75
     ]
   },
   {
@@ -599,7 +646,7 @@
       }
     ],
     "childNodeIndexes": [
-      73
+      78
     ]
   },
   {
@@ -607,8 +654,14 @@
     "nodeValue": "\n  "
   },
   {
+    "nodeType": 1,
+    "nodeName": "SCRIPT",
+    "attributes": [],
+    "childNodeIndexes": []
+  },
+  {
     "nodeType": 3,
-    "nodeValue": "\n\n"
+    "nodeValue": "\n\n\n"
   },
   {
     "nodeType": 1,
@@ -630,17 +683,13 @@
     "nodeName": "BODY",
     "attributes": [],
     "childNodeIndexes": [
-      17,
-      19,
-      20,
       22,
-      23,
+      24,
       25,
-      26,
+      27,
       28,
-      29,
+      30,
       31,
-      32,
       33,
       34,
       36,
@@ -651,12 +700,17 @@
       42,
       43,
       44,
-      71,
-      72,
-      74,
-      75,
+      46,
+      47,
+      48,
+      49,
       76,
-      77
+      77,
+      79,
+      80,
+      81,
+      82,
+      83
     ]
   },
   {
@@ -669,9 +723,9 @@
       }
     ],
     "childNodeIndexes": [
-      15,
-      16,
-      78
+      20,
+      21,
+      84
     ]
   }
 ]
